Bug 1530573 - Protect maybeResizeNursery() from overflow r=sfink
Differential Revision: https://phabricator.services.mozilla.com/D21158

UltraBlame original commit: a456a1594cb1960c465dc7a582134710f286fcf9

+++ js/src/gc/Nursery.h
@@ -639,8 +639,8 @@ class Nursery {
 
   
   void maybeResizeNursery(JS::GCReason reason);
-  void growAllocableSpace(unsigned newCapacity);
-  void shrinkAllocableSpace(unsigned newCapacity);
+  void growAllocableSpace(size_t newCapacity);
+  void shrinkAllocableSpace(size_t newCapacity);
   void minimizeAllocableSpace();
